+++ fast_dash/utils.py
@@ -11,44 +11,66 @@ import dash_bootstrap_components as dbc
 import dash_mantine_components as dmc
 from dash_iconify import DashIconify
 
+import docstring_parser
+import logging
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import PIL
+import pandas as pd
 import warnings
 import re
 
 
-def Fastify(
-    component, component_property, ack=None, placeholder=None, label_=None, tag=None
-):
+def Fastify(component, component_property, ack=None, placeholder=None, label_=None,tag=None, *args, **kwargs):
     """
-    Modify a Dash component to a FastComponent.
+    Modify a Dash component into a FastComponent.
 
     Args:
         component (Dash component): Dash component that needs to be modified
         component_property (str): Component property that's assigned the input or output values
         ack (Dash component, optional): Dash component that's displayed as an acknowledgement of the original component
-        placeholder (type, optional): Placeholder value of the component.
-        label_ (type, optional): Component title.
+        placeholder (str, optional): Placeholder value of the component.
+        label_ (str, optional): Component title.
+        tag (str, optional): Optional tag applied to the component.
 
     Returns:
         Fast component: Dash component modified to make it compatible with Fast Dash.
     """
+    
+    class FastComponent(type(component)):
+        def __init__(self, component, component_property, ack=ack, placeholder=placeholder, label_=label_, tag=tag, *args, **kwargs):
+            
+            self.component_property = component_property
+            self.ack = ack
+            self.label_ = label_
+            self.placeholder = placeholder
+            self.tag = tag
+            
+            # Copy normal attributes
+            for attr_name, attr_value in vars(component).items():
+                setattr(self, attr_name, attr_value)
 
-    component.component_property = component_property
-    component.ack = ack
-    component.label_ = label_
-    component.placeholder = placeholder
-    component.tag = tag
-
-    return component
+            # Copy the __doc__ attribute
+            self.__doc__ = component.__doc__
+            
+            for key, value in kwargs.items():
+                setattr(self, key, value)
+                
+        def __call__(self, **kwargs):            
+            self_copy = copy.deepcopy(self)
+            for key, value in kwargs.items():
+                setattr(self_copy, key, value)
+                
+            return self_copy
+        
+    return FastComponent(component, component_property, ack=ack, placeholder=placeholder, label_=label_, tag=tag)
 
 
 def Chatify(query_response_dict):
     "Convert a dictionary into a Chat component"
 
     if not isinstance(query_response_dict, dict):
-        raise TypeError("Chat component required a dictionary output.")
+        raise TypeError("Chat component requires a dictionary output ('query': ..., 'response': ...).")
 
     ## Chat component
     input_component = dbc.Row(
@@ -140,6 +162,7 @@ def theme_mapper(theme_name):
     return theme
 
 
+# Data type conversion utilities
 def _pil_to_b64(img):
     """
     Utility to convert PIL image to a base64 string.
@@ -203,6 +226,7 @@ def _get_input_names_from_callback_fn(callback_fn):
     return parameter_list
 
 
+# Fast Dash app utilities
 def _assign_ids_to_inputs(inputs, callback_fn):
     """
     Modify the 'id' property of inputs.
@@ -330,7 +354,9 @@ def _get_transform_function(output, tag):
         return Chatify
 
     subclass = type(output)
-    _transform_mapper = {plt.Figure: _mpl_to_b64, PIL.Image.Image: _pil_to_b64}
+    _transform_mapper = {plt.Figure: _mpl_to_b64, 
+                         PIL.Image.Image: _pil_to_b64,
+                         pd.DataFrame: lambda x: x.to_dict(orient="records")}
 
     for class_ in _transform_mapper:
         if issubclass(subclass, class_):
@@ -486,3 +512,70 @@ def _get_default_property(component_type):
         default_property = "value"
 
     return default_property
+
+
+def _parse_docstring_as_markdown(func, title=None, get_short=False):
+    """
+    Generate markdown documentation from the docstring of a function.
+
+    Args:
+        func (function): The function to document.
+        title (str, optional): Title of the text. If None, it is evaluated. Defaults to None.
+        get_short (bool, optional): If True, returns only the short description. Defaults to False.
+
+    Returns:
+        str: Markdown documentation string.
+    """
+    logging.warning("Parsing function docstring is still an experimental feature. To reduce uncertainty, consider setting `about` to `False`.")
+
+    parsed = docstring_parser.parse(func.__doc__)
+
+    if get_short == True:
+        return parsed.short_description
+
+    # Start with the function description
+    md_list = [
+        f"#### {func.__name__.replace('_', ' ').title() if not title else title}",
+        "",
+        parsed.short_description or "",
+        "",
+        parsed.long_description or "",
+        "",
+    ]
+
+    # Add parameters in table format
+    if parsed.params:
+        md_list.extend(
+            [
+                "##### Parameters",
+                "| Name | Type | Mandatory | Default | Description |",
+                "| ---- | ---- | --------- | ------- | ----------- |",
+            ]
+        )
+
+        for param in parsed.params:
+            param_line = (
+                f"| {param.arg_name} "
+                f"| {param.type_name or 'Not specified'} "
+                f"| {'No' if param.is_optional else 'Yes'} "
+                f"| {param.default or 'None'} "
+                f"| {param.description or 'No description provided'} |"
+            )
+            md_list.append(param_line)
+
+    md_list.extend(["# ", "# "])
+
+    # Add return values in table format
+    if parsed.returns:
+        md_list.extend(
+            ["", "##### Returns", "| Type | Description |", "| ---- | ----------- |"]
+        )
+
+        return_line = (
+            f"| {parsed.returns.type_name or 'Not specified'} "
+            f"| {parsed.returns.description or 'No description provided'} |"
+        )
+        md_list.append(return_line)
+
+    # Join all and return markdown string
+    return "\n".join(md_list)